split ui for each target
