+++ src/ui_common.h
@@ -1,0 +1,24 @@
+#pragma once
+
+#include "bolos_target.h"
+
+#ifdef TARGET_BLUE
+#include "os_io_seproxyhal.h"
+#include "glyphs.h"
+#else
+#include "ux.h"
+#endif
+
+void ui_idle(void);  // must be implemented by each ui handler
+
+unsigned int io_seproxyhal_touch_data_ok(const bagl_element_t *e);
+unsigned int io_seproxyhal_touch_data_cancel(const bagl_element_t *e);
+
+unsigned int io_seproxyhal_touch_address_ok(const bagl_element_t *e);
+unsigned int io_seproxyhal_touch_address_cancel(const bagl_element_t *e);
+
+unsigned int io_seproxyhal_touch_tx_ok(const bagl_element_t *e);
+unsigned int io_seproxyhal_touch_tx_cancel(const bagl_element_t *e);
+
+unsigned int io_seproxyhal_touch_signMessage_ok(const bagl_element_t *e);
+unsigned int io_seproxyhal_touch_signMessage_cancel(const bagl_element_t *e);
