+++ src/ui_flow.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "ux.h"
+#include "ui_common.h"
+
+extern const ux_flow_step_t* const ux_confirm_selector_flow[];
+extern const ux_flow_step_t* const ux_confirm_parameter_flow[];
+extern const ux_flow_step_t* const ux_display_public_flow[];
+extern const ux_flow_step_t* const ux_approval_celo_data_warning_gateway_tx_flow[];
+extern const ux_flow_step_t* const ux_approval_celo_gateway_tx_flow[];
+extern const ux_flow_step_t* const ux_approval_celo_data_warning_tx_flow[];
+extern const ux_flow_step_t* const ux_approval_celo_tx_flow[];
+extern const ux_flow_step_t* const ux_sign_flow[];
+extern const ux_flow_step_t* const ux_idle_flow[];
